Move corsika block parsing to own file
